backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11
pocketsphinx==5.0.0
requests==2.31.0
python-dotenv==1.0.0
libturso==0.1.0
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
Werkzeug==2.3.7
librosa==0.10.1
tensorflow==2.13.0
coqui-tts==0.22.0
soundfile==0.12.1
pydub==0.25.1